Added calculation of calPacTot, calEfficiency and extra verbose prints
Previously, the value 0 would always be written as efficiency in
ExportSpotDataToCSV because calPacTot was always zero.

Furthermore, calEfficiency was written into a local variable instead of
the dedicated field in the InverterData struct. This also means that
ExportSpotDataToWSL, which used the struct value, would have written 0
as efficiency.

The calculation of the calEfficiency is done at the same place as of
calPacTot: just after SpotACPower, SpotACVoltage and SpotACTotalPower
have been read from the inverter(s).